En casa sigo

+++ P4/delete.php
@@ -7,20 +7,26 @@
     <body>
     <?php
         require '.env.php';
+        require 'comprobacion.php';
+        if(auth()){
+            $db = new mysqli($SERVERNAME, $USERNAME, $PASSWORD, $DATABASE);
+            if($db->connect_error){
+                die('Connect Error (' . $db->connect_errno . ') ' . $db->connect_error);
+            }
 
-        $db = new mysqli($SERVERNAME, $USERNAME, $PASSWORD, $DATABASE);
-        if($db->connect_error){
-            die('Connect Error (' . $db->connect_errno . ') ' . $db->connect_error);
-        }
-
-        $query = "DELETE FROM EMPLEADOS WHERE NOMBRE LIKE " . "'" . $_GET['NOMBRE'] . "';";
-        if(!$db->query($query)){
-            die('Connect Error (' . $db->connect_errno . ') ' . $db->connect_error);
+            $query = "DELETE FROM EMPLEADOS WHERE NOMBRE LIKE " . "'" . $_GET['NOMBRE'] . "';";
+            if(!$db->query($query)){
+                die('Connect Error (' . $db->connect_errno . ') ' . $db->connect_error);
+            }
+            else{
+                $db->close();
+                header("Location: index.php");
+                exit;
+            }
         }
         else{
-            $db->close();
-            header("Location: index.php");
-            exit;
+            echo '<h1>401 Unauthorized</h1>';
+            header("HTTP/1.0 401 Unauthorized");
         }
     ?>
     </body>
